Fixed hal-hal yang ada di codeReview: - ilangin static func yang ada di categoryChart - Fixed bug di summaryView
